requirements.txt
flask
flask_sqlalchemy
werkzeug
nanoid
Flask-Migrate
psycopg2
python-dotenv
sqlalchemy_utils